+++ src/app/public/modules/tabs/tabset.component.ts
@@ -1,4 +1,8 @@
 import {
+  Location
+} from '@angular/common';
+
+import {
   AfterContentInit,
   AfterViewInit,
   ChangeDetectorRef,
@@ -9,6 +13,7 @@ import {
   Input,
   OnChanges,
   OnDestroy,
+  Optional,
   Output,
   QueryList,
   SimpleChanges
@@ -16,9 +21,12 @@ import {
 
 import {
   ActivatedRoute,
-  Params,
   Router
 } from '@angular/router';
+
+import {
+  Observable
+} from 'rxjs/Observable';
 
 import {
   Subject
@@ -39,6 +47,10 @@ import {
 } from './tabset-adapter.service';
 
 import {
+  SkyTabsetPermalinkParams
+} from './tabset-permalink-params';
+
+import {
   SkyTabsetService
 } from './tabset.service';
 
@@ -52,7 +64,7 @@ import {
   ]
 })
 export class SkyTabsetComponent
-  implements AfterContentInit, AfterViewInit, OnDestroy, OnChanges {
+  implements AfterContentInit, AfterViewInit, OnChanges, OnDestroy {
 
   /**
    * Specifies the index of the active tab.
@@ -132,6 +144,8 @@ export class SkyTabsetComponent
   @ContentChildren(SkyTabComponent)
   public tabs: QueryList<SkyTabComponent>;
 
+  private activeIndexOnLoad: number | string;
+
   private ngUnsubscribe = new Subject<void>();
 
   private _permalinkId: string;
@@ -144,7 +158,8 @@ export class SkyTabsetComponent
     private elRef: ElementRef,
     private changeRef: ChangeDetectorRef,
     private activatedRoute: ActivatedRoute,
-    private router: Router
+    private location: Location,
+    @Optional() private router?: Router
   ) { }
 
   public getTabButtonId(tab: SkyTabComponent): string {
@@ -173,7 +188,7 @@ export class SkyTabsetComponent
 
   public selectTab(tab: SkyTabComponent): void {
     if (this.permalinkId && tab.permalinkValue) {
-      return;
+      this.setPathParamPermalinkValue(tab.permalinkValue);
     }
 
     this.tabsetService.activateTab(tab);
@@ -190,9 +205,6 @@ export class SkyTabsetComponent
   }
 
   public ngAfterContentInit(): void {
-    // Initialize each tab's index (in case tabs are instantiated out of order).
-    this.tabs.forEach(tab => tab.initializeTabIndex());
-
     this.tabs.changes
       .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe((change: QueryList<SkyTabComponent>) => {
@@ -209,8 +221,16 @@ export class SkyTabsetComponent
       });
 
     if (this.active !== undefined) {
+      this.activeIndexOnLoad = this.active;
       this.tabsetService.activateTabIndex(this.active);
     }
+
+    // Render the template before activating a tab.
+    setTimeout(() => {
+      // Initialize each tab's index (in case tabs are instantiated out of order).
+      this.tabs.forEach(tab => tab.initializeTabIndex());
+      this.activateTabByPermalinkValue();
+    });
 
     this.tabsetService.activeIndex
       .pipe(
@@ -223,16 +243,20 @@ export class SkyTabsetComponent
         setTimeout(() => {
           if (newActiveIndex !== this.active) {
             this.active = newActiveIndex;
+            if (this.activeIndexOnLoad === undefined) {
+              this.activeIndexOnLoad = newActiveIndex;
+            }
             this.activeChange.emit(newActiveIndex);
           }
         });
       });
 
-    // Wait for the tab components' `active` state to be resolved before
-    // listening to changes to the URL params.
-    setTimeout(() => {
-      this.watchQueryParamChanges();
-    });
+      // Listen for back/forward history button presses to detect path param changes in the URL.
+      // (Angular's router events observable doesn't emit when path params change.)
+      // See: https://stackoverflow.com/a/51471155/6178885
+      Observable.fromEvent(window, 'popstate')
+        .takeUntil(this.ngUnsubscribe)
+        .subscribe(() => this.activateTabByPermalinkValue());
   }
 
   public ngAfterViewInit(): void {
@@ -244,6 +268,7 @@ export class SkyTabsetComponent
         this.updateDisplayMode(currentOverflow);
       });
 
+    // Render the template before setting display mode.
     setTimeout(() => {
       this.adapterService.detectOverflow();
       this.updateDisplayMode(this.adapterService.currentOverflow);
@@ -255,7 +280,20 @@ export class SkyTabsetComponent
     this.ngUnsubscribe.next();
     this.ngUnsubscribe.complete();
     /*tslint:disable-next-line:no-null-keyword*/
-    this.setQueryParamPermalinkValue(null);
+    this.setPathParamPermalinkValue(null);
+  }
+
+  public getPathParams(): SkyTabsetPermalinkParams {
+    const params: SkyTabsetPermalinkParams = {};
+
+    const existingParamPairs = this.location.path().split(';');
+    existingParamPairs.shift();
+    existingParamPairs.forEach((pair) => {
+      const fragments = pair.split('=');
+      params[fragments[0]] = fragments[1];
+    });
+
+    return params;
   }
 
   private updateDisplayMode(currentOverflow: boolean): void {
@@ -263,41 +301,24 @@ export class SkyTabsetComponent
     this.changeRef.markForCheck();
   }
 
-  private watchQueryParamChanges(): void {
-    this.activatedRoute.queryParams
-      .pipe(
-        distinctUntilChanged(),
-        takeUntil(this.ngUnsubscribe)
-      )
-      .subscribe((params) => {
-        if (!this.permalinkId) {
-          return;
-        }
+  private activateTabByPermalinkValue(): void {
+    const params = this.getPathParams();
 
-        const permalinkValue = params[this.permalinkId];
-        if (permalinkValue) {
-          this.activateTabByPermalinkValue(permalinkValue);
-        } else {
-          this.setQueryParamByActiveTab();
-        }
-      });
-  }
+    if (
+      !(this.permalinkId in params) &&
+      this.activeIndexOnLoad !== undefined
+    ) {
+      this.tabsetService.activateTabIndex(this.activeIndexOnLoad);
+      return;
+    }
 
-  private setQueryParamByActiveTab(): void {
-    this.tabsetService.tabs
-      .pipe(take(1))
-      .subscribe((tabs) => {
-        const activeTab = tabs.find(tab => tab.active);
-        this.setQueryParamPermalinkValue(activeTab.permalinkValue);
-      });
-  }
+    const value = params[this.permalinkId];
 
-  private activateTabByPermalinkValue(value: string): void {
-    let index: number;
+    let index: number | string;
 
     this.tabs.forEach((tabComponent, i) => {
       if (tabComponent.permalinkValue === value) {
-        index = i;
+        index = tabComponent.tabIndex;
       }
     });
 
@@ -307,15 +328,19 @@ export class SkyTabsetComponent
     }
   }
 
-  private setQueryParamPermalinkValue(value: string): void {
+  private setPathParamPermalinkValue(value: string): void {
     if (this.permalinkId) {
-      const queryParams: Params = {};
-      queryParams[this.permalinkId] = value;
+      const params = this.getPathParams();
 
-      this.router.navigate([], {
-        queryParams,
-        queryParamsHandling: 'merge'
-      });
+      params[this.permalinkId] = value;
+
+      // Update the URL without triggering a navigation state change.
+      // See: https://stackoverflow.com/a/46486677
+      const url = this.router.createUrlTree([params], {
+        relativeTo: this.activatedRoute
+      }).toString();
+
+      this.location.go(url);
     }
   }
 }